# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FinMatrix Financial Intelligence Dashboard API
 * OpenAPI spec version: 1.0.0
 */

export interface HealthStatus {
  status: string;
}